% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fasano_2D_dvals.R
\name{Fasano_2D_dvals}
\alias{Fasano_2D_dvals}
\title{Calculate D-values for a Two-dimensional Kolmogorov-Smirnov Test}
\usage{
Fasano_2D_dvals(xy)
}
\arguments{
\item{xy}{The D-values for each observation as a vector.}
}
\value{
The original data frame with a new "maxdiff" column containing D-values.
}
\description{
Find D-values of observations for a two-dimensional Kolmogorov-Smirnov test as defined by [Fasano and Franceschini 1987](https://academic.oup.com/mnras/article/225/1/155/1007281).
}
