% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fasano_3D_dvals.R
\name{Fasano_3D_dvals}
\alias{Fasano_3D_dvals}
\title{Calculate D-values for a Three-dimensional Kolmogorov-Smirnov Test}
\usage{
Fasano_3D_dvals(xcol, ycol, zcol)
}
\arguments{
\item{xcol}{a vector of x values}

\item{ycol}{a vector of y values}

\item{zcol}{a vector of z values}
}
\value{
The D-values for each set of observations as a vector
}
\description{
Find D-values of observations for a three-dimensional Kolmogorov-Smirnov test as defined by Fasano and Franceschini (1987).
}
\examples{
xcol <- rnorm(10)
ycol <- rnorm(10)
zcol <- rnorm(10)

Fasano_3D_dvals(xcol, ycol, zcol)

\dontrun{
xcol <- rnorm(10)
ycol <- rnorm(10)
zcol <- rnorm(5)

Fasano_3D_dvals(xcol, ycol, zcol)
}
}
