% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fasano_2D_dvals.R
\name{Fasano_2D_dvals}
\alias{Fasano_2D_dvals}
\title{Calculate D-values for Kolmogorov-Smirnov Test}
\usage{
Fasano_2D_dvals(xy)
}
\arguments{
\item{xy}{The D-values for each observation as a vector.}
}
\value{
The original data frame with a new "maxdiff" column containing D-values.
}
\description{
Find number of observations in each quadrant of a dataset needed to run a two-dimensional Kolmogorov-Smirnov test as defined by [https://watermark.silverchair.com/mnras225-0155.pdf?token=AQECAHi208BE49Ooan9kkhW_Ercy7Dm3ZL_9Cf3qfKAc485ysgAAAs0wggLJBgkqhkiG9w0BBwagggK6MIICtgIBADCCAq8GCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMgHp1AmsYKnHXNEOdAgEQgIICgM8aZSoVKTeXKk_3u2Q16WV38bdHKZl47PPAZRnydI6VywiYEDK-o9G4LCYrsBtp9W7CQdG-sCouNnAg31KLOaX34sD0cYkMKIc0SELP0rLEkJIMTSsMBRdyn01EmO5hrH9_kGe24VPb9RIQq2LEB31sbzQmAOzwnGAOaPD7rlw0hn8cXxMgYjOpUDkJE9SDb5fhMmrZSOodnBcRAP3pdc_egvoYTnAmUqDSI-5AKHj9pErWPhVrAJ0HkdNTfI1CQLyXAqDSgoD8l109TgAfQAKk7WZ-Z0oP_wNT4ClVKV6cbMEIX1Hszd1JNShLs6qYg38RrZAi8lEAiONfvRnfBFdS7i_easW9jBrt0QWTcno-QopLWUXze6rE7qvyFjiqyB3g-UvKv0P0HOv4DgYw8STmvIZ_0KDh5YMn3y6uPG2nvP8ktudoaY1wZ2WZwE_IoxLRt6WZ0G1o9FOmnLw_tXRosMGKnCgV1YKSbshPC5rKZ8aoHOsmhhUfRxzwLv6ew9Am8M_2wINVaBn9HnGbu8WVRzRjvOlXTEaNk-ZUbTVgsPxCU2H4isCjrWbHXoCVGf4r8lM3a6Pc6oqXl3fIvZwvQmgPt3KgwFZxyqJHsXMJZWXPAZrU8c81EzdQ4816AyRGo8p0MlWmk7ArmBJ_uaOGu3jMGXnLJzJrzTZ3B-nPiLP7ctrfZsFcWipRuhGlWtL9PcKK3FBWa_gSprHUtEW_VHLihVHJOSU59qwoBao2sxsGQbI-3fKVSDofPsn-qgKKgPWtyM_K6CA_fDFjxkyUy7fy1hIm-ix50f9pMVsA8HpSPz4CDC7qywL5yrG8a3AeHzq8H8vesd33rfTuKDg](Fasano and Franceschini 1987).
}
