% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fasano_2D_dvals.R
\name{Fasano_2D_dvals}
\alias{Fasano_2D_dvals}
\title{Calculate D-values for a Two-dimensional Kolmogorov-Smirnov Test}
\usage{
Fasano_2D_dvals(xcol, ycol)
}
\arguments{
\item{xcol}{a vector of x values}

\item{ycol}{a vector of y values}
}
\value{
The D-values for each pair of observations as a vector
}
\description{
Find D-values of observations for a two-dimensional Kolmogorov-Smirnov test as defined by Fasano and Franceschini (1987).
}
\examples{
xcol <- rnorm(10)
ycol <- rnorm(10)

Fasano_2D_dvals(xcol, ycol)

\dontrun{
xcol <- rnorm(10)
ycol <- rnorm(5)

Fasano_2D_dvals(xcol, ycol)
}
}
