% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fasano_2DKS.R
\name{Fasano_2DKS}
\alias{Fasano_2DKS}
\title{Run a Two-dimensional Kolmogorov-Smirnov Test}
\usage{
Fasano_2DKS(xcol, ycol, rands = 5000, alpha = 0.05)
}
\arguments{
\item{xcol}{a vector of x values}

\item{ycol}{a vector of y values}

\item{rands}{A numeric indicating how many randomizations to perform. Default is 5000.}

\item{alpha}{A numeric between 0 and 1 for the desired alpha level. Default is 0.05.}
}
\value{
A data frame of summary values and statistics (the maximum D-value; the x, y, and p values at the maximum D-value; and the minimum and maximum x and y values for the statistically-significant observations)
}
\description{
Run a two-dimensional Kolmogorov-Smirnov test as defined by \href{https://academic.oup.com/mnras/article/225/1/155/1007281}{Fasano and Franceschini 1987}. Allows for specification of the number of randomization to perform and the desired alpha level.
}
\examples{
xcol <- rnorm(10)
ycol <- rnorm(10)

Fasano_2DKS(xcol, ycol, rands = 2000, alpha = 0.05)

\dontrun{
xcol <- rnorm(10)
ycol <- rnorm(5)

Fasano_3DKS(xcol, ycol, rands = 2000, alpha = 0.05)
}
}
